GUA-209: Refactor uploading check and metadata
Add a helper function to upload the JWT and its metadata, and move
functions that write files to the pod out to `lib/pod.ts`
